requirements.txt
opencv-python>=4.5.0
numpy>=1.21.0
pyautogui>=0.9.53
keyboard>=0.13.5
Pillow>=9.0.0
pytesseract>=0.3.10
